% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_array.R
\name{anno_expr}
\alias{anno_expr}
\title{Annotating array probe expression matrix}
\usage{
anno_expr(
  arrayMat,
  anno,
  probe_col = NULL,
  symbol = "symbol",
  probe = "probe_id"
)
}
\arguments{
\item{arrayMat}{a data frame with probe id as first column and other columns are expression of gene in each samples}

\item{anno}{annotation file with gene symbol and other ids matched to probe id of \code{arrayMat},
at least contain two column of probe id and symbol id.}

\item{probe_col}{The column name or number of probe id in \code{arrayMat}. Dafult is NULL, choose the first column as probe id.}

\item{symbol}{column name of gene symbol in annotation file}

\item{probe}{column name of relevant id in annotation file}
}
\value{
return a gene expression matrix
}
\description{
Annotating array probe expression matrix and remove duplicated genes
}
