% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anno_hug133plus2}
\alias{anno_hug133plus2}
\title{the annotation file of hug133plus2 array}
\format{
\describe{
\item{probe_id}{array probe id}
\item{symbol}{probe coresponding gene id}
}
}
\usage{
anno_hug133plus2
}
\description{
A dataset containing the probe id and coresponding gene id of hug133plus2 array
}
\keyword{datasets}
