% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_expr.R
\name{check_expr}
\alias{check_expr}
\title{Check the data quality of expression matrix}
\usage{
check_expr(data, id_col = NULL, verbose = TRUE)
}
\arguments{
\item{data}{a gene or probe data frame like \code{arrayMat} in \code{\link{anno_expr}} or \code{exprMat} in \code{\link{remove_duplicate}}}

\item{id_col}{The column name or number of geneid or probe id. Dafult is NULL, choose the first column as id.}

\item{verbose}{output other useful information, default TRUE.}
}
\value{
return a logical value.
}
\description{
Check the expression matrix whether need other process
}
\details{
The expression matrix will be checked for three task:
\itemize{
\item check whether have NA
\item check whether have Inf or -Inf
\item check the standard deviation(sd) whether is 0
\item check whether have duplicated genes
}
if have problem, return FALSE, else no problem return TRUE.
}
