% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_expr.R
\name{convert_expr}
\alias{convert_expr}
\title{Convert read counts to transcripts per million (TPM)}
\usage{
convert_expr(
  exprMat,
  idType = c("ensembl", "entrez", "symbol"),
  gene_col = NULL,
  from = "count",
  to = "tpm",
  org = c("human", "mouse"),
  effLength = NULL
)
}
\arguments{
\item{exprMat}{A data frame with geneid as as first column and other columns are expression of gene in each samples.}

\item{idType}{Type of gene id. "ensembl" "entrez","symbol"}

\item{gene_col}{The column name or number of geneid in \code{exprMat}. Dafult is NULL, choose the first column as geneid.}

\item{from}{The type of matrix, support: "count", "tpm", "fpkm", "rpkm"}

\item{to}{Which type will be converted to, support: "tpm", "fpkm", "rpkm"}

\item{org}{Organism type, support human and mouse. We used the gen length information get from ensembl website (www.ensembl.org),
if you use the custom gene length information, you can ignore it.}

\item{effLength}{The gene length information which contains gene id \code{id} and gene length \code{length}. Dafult is NULL, we use the
information stored in this package.}
}
\description{
Convert the count matrix to TPM matrix
}
