% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human}
\alias{human}
\title{gene length information of human}
\format{
An object of class \code{data.frame} with 67490 rows and 6 columns.
}
\source{
\url{www.ensembl.org}
}
\usage{
human
}
\description{
The gene length and id information of human acessed from ensembl
}
\keyword{datasets}
