% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse}
\alias{mouse}
\title{gene length information of mouse}
\format{
An object of class \code{data.frame} with 53982 rows and 6 columns.
}
\source{
\url{www.ensembl.org}
}
\usage{
mouse
}
\description{
The gene length and id information of mouse acessed from ensembl
}
\keyword{datasets}
