% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicate.R
\name{remove_duplicate}
\alias{remove_duplicate}
\title{Remove duplicate gene symbol}
\usage{
remove_duplicate(
  exprMat,
  gene_col = NULL,
  method = c("order", "combine"),
  value = c("mean", "median")
)
}
\arguments{
\item{exprMat}{A data frame with geneid as as first column and other columns are expression of gene in each samples.}

\item{gene_col}{The column name or number of geneid in \code{exprMat}. Dafult is NULL, choose the first column as geneid.}

\item{method}{method used to filter duplicate genes; filter duplicated gene by the mean of median value, or combine the value by mean or median value instead}

\item{value}{use mean of median value as reference}
}
\value{
return a expression matrix
}
\description{
Remove duplicate gene symbol on a gene expression data
}
\details{
We use the mean or median value of the duplicated genes as the expression value of this gene.
}
