% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_data.R
\name{transfer_data}
\alias{transfer_data}
\title{Transform the expression matrix to normal quality}
\usage{
transfer_data(
  exprMat,
  gene_col = NULL,
  data_type = c("all", "NA", "Inf"),
  data_to = 0
)
}
\arguments{
\item{exprMat}{a gene or probe data frame like \code{arrayMat} in \code{\link{anno_expr}} or \code{exprMat} in \code{\link{remove_duplicate}}}

\item{gene_col}{The column name or number of geneid in \code{exprMat}. Dafult is NULL, choose the first column as geneid.}

\item{data_type}{the data type to change, NA, Inf or all}

\item{data_to}{transform data into 0 or 1, or others, default is 0}
}
\value{
return minipulated data
}
\description{
Transform the expression matrix  with bad quality to normal quality
}
