% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log2expr.R
\name{log2expr}
\alias{log2expr}
\title{log2 transformation of expression matrix}
\usage{
log2expr(exprMat)
}
\arguments{
\item{exprMat}{gene expression matrix with row as genes and sample in the column}
}
\value{
a gene expression matrix after log2 transformation
}
\description{
Do the expression matrix needs log2 transformation
}
\details{
The function will first check whether the expression matrix has undergone
log2 transformation; if the expression matrix have done log2 transformation, return
the raw expression matrix, else do the log2 transformation
}
\examples{
  mat <- matrix(sample(1:10, 30, replace = T), ncol = 5)
  log2_mat <- log2expr(exprMat = mat)
  res <- log2expr(exprMat = log2_mat)
}
